pyproject.toml
[project]
name = "ragstage2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.20.7",
    "datasets>=4.8.3",
    "google-generativeai>=0.8.6",
    "langchain-chroma>=1.1.0",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "pandas>=2.3.3",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tenacity>=9.1.4",
]
